google_calendar/requirements.txt
google-search-results==2.4.2
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.6
pytest==7.4.0
pytest-mock==3.11.1
typing_extensions==4.7.1
urllib3==1.26.16
python-dateutil==2.8.2
